use correct voter token record when relinquishing veto (#1725)
