+++ components/VotePanel/hooks.ts
@@ -1,4 +1,5 @@
 import { useRouteProposalQuery } from '@hooks/queries/proposal'
+import { useRealmQuery } from '@hooks/queries/realm'
 import {
   useUserCommunityTokenOwnerRecord,
   useUserCouncilTokenOwnerRecord,
@@ -6,8 +7,14 @@ import {
 import useRoleOfGovToken from '@hooks/selectedRealm/useRoleOfToken'
 import { useHasVoteTimeExpired } from '@hooks/useHasVoteTimeExpired'
 import { useProposalGovernanceQuery } from '@hooks/useProposal'
-import { ProposalState, Proposal, Governance } from '@solana/spl-governance'
+import {
+  ProposalState,
+  Proposal,
+  Governance,
+  VoteThresholdType,
+} from '@solana/spl-governance'
 import dayjs from 'dayjs'
+import { useMemo } from 'react'
 
 export const useIsVoting = () => {
   const proposal = useRouteProposalQuery().data?.result
@@ -66,5 +73,38 @@ export const useVoterTokenRecord = () => {
 
   const voterTokenRecord =
     votingPop === 'community' ? ownTokenRecord : ownCouncilTokenRecord
+  return voterTokenRecord
+}
+
+/*
+  returns: undefined if loading, false if nobody can veto, 'council' if council can veto, 'community' if community can veto
+*/
+export const useVetoingPop = () => {
+  const tokenRole = useVotingPop()
+  const governance = useProposalGovernanceQuery().data?.result
+  const realm = useRealmQuery().data?.result
+  const vetoingPop = useMemo(() => {
+    if (governance === undefined) return undefined
+
+    return tokenRole === 'community'
+      ? governance?.account.config.councilVetoVoteThreshold.type !==
+          VoteThresholdType.Disabled &&
+          // if there is no council then there's not actually a vetoing population, in my opinion
+          realm?.account.config.councilMint !== undefined &&
+          'council'
+      : governance?.account.config.communityVetoVoteThreshold.type !==
+          VoteThresholdType.Disabled && 'community'
+  }, [governance, tokenRole, realm?.account.config.councilMint])
+
+  return vetoingPop
+}
+
+export const useUserVetoTokenRecord = () => {
+  const ownTokenRecord = useUserCommunityTokenOwnerRecord().data?.result
+  const ownCouncilTokenRecord = useUserCouncilTokenOwnerRecord().data?.result
+
+  const vetoingPop = useVetoingPop()
+  const voterTokenRecord =
+    vetoingPop === 'community' ? ownTokenRecord : ownCouncilTokenRecord
   return voterTokenRecord
 }
